# src/main/resources/api.yml
openapi: 3.0.1
info:
  title: Demo
  description: 'Demo'
  version: 0.5.0
servers:
  - url: https://op.op.com
...
paths:
...
  /asistencia:
    get:
      tags:
        - Asistencia
      summary: Retorna todas las asistencias
      operationId: getAllAsistencias
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsistenciasResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    post:
      tags:
        - Asistencia
      summary: Crea una asistencia
      operationId: saveAsistencia
      requestBody:
        description: Crea una nueva asistencia
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsistenciaRequest'
        required: true
      responses:
        200:
          description: Asistencia creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsistenciaResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /asistencia/tecnico/{idTechnician}:
    get:
      tags:
        - Asistencia
      summary: Retorna todas las asistencias
      operationId: getAsistenciasByTechnician
      parameters:
        - name: idTechnician
          in: path
          description: ID del tecnico encargado
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Operacion Exitosa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsistenciasResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
  /asistencia/{id}:
    get:
      tags:
        - Asistencia
      summary: Retorna una Asistencia especifica
      operationId: getAsistenciaById
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a retornar
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Asistencia encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsistenciaResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    put:
      tags:
        - Asistencia
      summary: Actualizar una Asistencia
      operationId: updateAsistencia
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar una asistencia existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsistenciaRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
    patch:
      tags:
        - Asistencia
      summary: Actualizar el estado de una Asistencia
      operationId: updateEstadoAsistencia
      parameters:
        - name: id
          in: path
          description: ID de la asistencia a editar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Actualizar el estado de una asistencia existente
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EstadoAsistenciaRequest'
        required: true
      responses:
        200:
          description: Asistencia actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        401:
          description: No Autorizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        404:
          description: No se encontro el reporte
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
        500:
          description: Error en el servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'
...
components:
  schemas:
...
    AsistenciaRequest:
      type: object
      required:
        - nombre
        - apellido
        - telefono
        - tipoSituacion
        - placa
        - latitud
        - longitud
      properties:
        fechaAsistencia:
          type: string
          format: date-time
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        tipoSituacion:
          type: string
          maxLength: 100
        situacion:
          type: string
          maxLength: 2000
        placa:
          type: string
          maxLength: 20
        tipoVehiculo:
          type: string
          maxLength: 50
        marca:
          type: string
          maxLength: 50
        modelo:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 20
        urlImagenFirma:
          type: string
          maxLength: 250
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
        idTechnician:
          type: integer
          format: int64
        notasTechnician:
          type: string
          maxLength: 2000
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    EstadoAsistenciaRequest:
      type: object
      required:
        - estado
        - notasTechnician
      properties:
        notasTechnician:
          type: string
          maxLength: 2000
        urlImagenFirma:
          type: string
          maxLength: 250
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    Asistencia:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fechaLlamada:
          type: string
          format: date-time
        fechaAsistencia:
          type: string
          format: date-time
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        tipoSituacion:
          type: string
          maxLength: 100
        situacion:
          type: string
          maxLength: 2000
        placa:
          type: string
          maxLength: 20
        tipoVehiculo:
          type: string
          maxLength: 50
        marca:
          type: string
          maxLength: 50
        modelo:
          type: string
          maxLength: 50
        color:
          type: string
          maxLength: 20
        urlImagenFirma:
          type: string
          maxLength: 250
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
        idTechnician:
          type: integer
          format: int64
        notasTechnician:
          type: string
          maxLength: 2000
        tecnico:
          $ref: '#/components/schemas/Technician'
        estado:
          type: string
          description: Estado de la asistencia
          enum:
            - Enviada
            - Asignada
            - En Proceso
            - Completada
            - Cerrada
    AsistenciaResponse:
      type: object
      properties:
        value:
          $ref: '#/components/schemas/Asistencia'
        type:
          type: string
        message:
          type: string
    AsistenciasResponse:
      type: object
      properties:
        value:
          type: array
          items:
            $ref: '#/components/schemas/Asistencia'
        type:
          type: string
        message:
          type: string
...
    Technician:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
          maxLength: 100
        apellido:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
        latitud:
          type: string
          maxLength: 50
        longitud:
          type: string
          maxLength: 50
...
    EmptyResponse:
      type: object
      properties:
        type:
          type: string
        message:
          type: string